Optimize parameter passing, make method inline, optimize uint sizes, and refactor validate_perpendicular_rows to only check the housing perpendicular row.

+++ include/dictionary.h
@@ -14,7 +14,7 @@ typedef struct {
 
 typedef struct {
     const char *word;
-    uint16_t word_length;
+    uint8_t word_length;
     Row rows[BOARD_SIZE];
     uint16_t valid_starts;
 } WordStartRowEntry;
